+++ cookbooks/aws-parallelcluster-environment/resources/efs/partial/_common.rb
@@ -6,8 +6,20 @@ def _efs_utils_version
   node['cluster']['efs']['version']
 end
 
-def _efs_utils_checksum
-  node['cluster']['efs']['sha256']
+# Tracked major line; the repo installs the newest release below the next major
+# (EFS guarantees no breaking changes within a major).
+def _efs_utils_major
+  _efs_utils_version.split('.').first.to_i
+end
+
+# Public EFS package endpoint (CloudFront).
+def efs_domain
+  "https://amazon-efs-utils.aws.com"
+end
+
+# DevSetting: skip the efs-utils install.
+def _skip_efs_utils_install?
+  node['cluster']['efs']['skip_install'].to_s == 'true'
 end
 
 def already_installed?(package_name, expected_version)
@@ -30,5 +42,14 @@ action :increase_poll_interval do
     pattern "poll_interval_sec = 1$"
     line "poll_interval_sec = 10"
     replace_only true
+  end
+end
+
+action :install_efs_utils_within_major do
+  # `package` can't express a version range, so cap the major via dnf.
+  execute "install amazon-efs-utils" do
+    command "dnf install -y 'amazon-efs-utils < #{_efs_utils_major + 1}'"
+    retries 3
+    retry_delay 5
   end
 end